requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
typer>=0.9.0
click>=8.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Async
anyio>=4.0.0

# Data
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# AI/ML & LangChain/LangGraph
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
tiktoken>=0.5.0

# Vector Stores
chromadb>=0.4.0
weaviate-client>=3.25.0
faiss-cpu>=1.7.4

# Distributed
ray[default,serve]>=2.9.0

# Web
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing & Dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
